requirements.txt
boto3==1.9.212
botocore==1.12.212
Django==2.2.4
django-cors-headers==3.1.0
django-redis==4.10.0
djangorestframework==3.10.2
docutils==0.15.2
jmespath==0.9.4
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
redis==3.3.8
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
